Changed to statements_extra
this is a single table which includes any data relevant to a specific statement, e.g. when viewing the coding tab. So there is a lot of duplication as all the relevant respondent info is available for every statement. But it is flexible because any number of nested units possible